Main page icons ui fixed
